Feature: dhcp for eth0

+++ installers/uninstall.sh
@@ -18,8 +18,10 @@ set -o errtrace
 readonly raspap_dir="/etc/raspap"
 readonly raspap_user="www-data"
 readonly raspap_sudoers="/etc/sudoers.d/090_raspap"
-readonly raspap_dnsmasq="/etc/dnsmasq.d/090_raspap.conf"
+readonly raspap_default="/etc/dnsmasq.d/090_raspap.conf"
+readonly raspap_wlan0="/etc/dnsmasq.d/090_wlan0.conf"
 readonly raspap_sysctl="/etc/sysctl.d/90_raspap.conf"
+readonly raspap_adblock="/etc/dnsmasq.d/090_adblock.conf"
 readonly raspap_network="/etc/systemd/network/"
 readonly rulesv4="/etc/iptables/rules.v4"
 webroot_dir="/var/www/html"
@@ -139,7 +141,9 @@ function _remove_raspap_directories() {
 # Removes raspapd.service
 function _remove_raspap_service() {
     _install_log "Removing raspapd.service"
-    sudo rm /lib/systemd/system/raspapd.service || _install_error "Unable to remove raspap.service file"
+    if [ -f /lib/systemd/system/raspapd.service ]; then
+        sudo rm /lib/systemd/system/raspapd.service || _install_error "Unable to remove raspap.service file"
+    fi
     sudo systemctl daemon-reload
     sudo systemctl disable raspapd.service || _install_error "Failed to disable raspap.service"
     echo "Done."
@@ -172,9 +176,19 @@ function _restore_networking() {
     echo "Done."
     # Remove dnsmasq and bridge configs
     echo "Removing 090_raspap.conf from dnsmasq"
-    sudo rm "$raspap_dnsmasq" || _install_error "Unable to remove $raspap_dnsmasq"
+    if [ -f $raspap_default ]; then
+        sudo rm "$raspap_default" || _install_error "Unable to remove $raspap_default"
+    fi
+    echo "Removing 090_wlan0.conf from dnsmasq"
+    if [ -f $raspap_wlan0 ]; then
+        sudo rm "$raspap_wlan0" || _install_error "Unable to remove $raspap_wlan0"
+    fi
     echo "Removing raspap bridge configurations"
     sudo rm "$raspap_network"/raspap* || _install_error "Unable to remove bridge config"
+    if [ -f $raspap_adblock ]; then
+        echo "Removing raspap adblock configuration"
+        sudo rm "$raspap_adblock" || _install_error "Unable to remove adblock config"
+    fi
 }
 
 # Removes installed packages
